use same logic to format durations of an album and a scrobble
