+++ src/domains/scrobbleAlbum/partials/Tracklist.tsx
@@ -23,18 +23,9 @@ import type { Album, DiscogsAlbum } from 'utils/types/album';
 import type { Scrobble } from 'utils/types/scrobble';
 import type { Track } from 'utils/types/track';
 
-const DateTimePicker = lazyWithPreload(() => import('components/DateTimePicker'));
+import { formatDuration } from 'utils/datetime'
 
-function formatDuration(totalSeconds: number) {
-  const datetime = addSeconds(new Date(0), totalSeconds);
-  const h = datetime.getUTCHours();
-  const m = datetime.getUTCMinutes();
-  const s = datetime.getUTCSeconds();
-  
-  const mm = (m < 10) ? ("0" + m) : m.toString();
-  const ss = (s < 10) ? ("0" + s) : s.toString();
-  return (h > 0) ? `${h}:${mm}:${ss}` : `${mm}:${ss}`;
-}
+const DateTimePicker = lazyWithPreload(() => import('components/DateTimePicker'));
 
 // ToDo: refactor this component completely.
 // It's too complex and carries several blocks from old code.
